Move database calls to own folder
